--- a/pkg/infoschema/tables.go
+++ b/pkg/infoschema/tables.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"net"
 	"net/http"
+	"runtime"
 	"slices"
 	"sort"
 	"strconv"
@@ -1982,13 +1983,19 @@ func GetClusterServerInfo(ctx sessionctx.Context) ([]ServerInfo, error) {
 			return nil, err
 		}
 
+		//Limit concurrency to number of CPU cores
+		concurrencyLimit := runtime.GOMAXPROCS(0)
+		sem := make(chan struct{}, concurrencyLimit)
+
 		//concurrently resolve addresses
 		var wg sync.WaitGroup
 		for i := range nodes {
 			wg.Add(1)
 			node := &nodes[i] //copy pointer for goroutine
+			sem <- struct{}{} //acquire slot
 			go func(n *ServerInfo) {
 				defer wg.Done()
+				defer func() { <-sem }() //release slot
 				n.ResolveLoopBackAddr()
 			}(node)
 		}