@hawkingrei 2025-06-06
/hold

If there are 1,000 nodes, would you just launch 1,000 goroutines to handle them? I think we need to add a limit on concurrency; otherwise, it's just too risky.