@abrarr21 2025-06-06
@hawkingrei That's a valid point. Would it make sense to use a buffered channel (as a semaphore) to limit the number of concurrent goroutines? For example, we could cap it at 32 or 64.

Let me know your thoughts — happy to implement it accordingly.